notebooks/pathway-demo/requirements.txt
aneris-iamc==0.3.1
anyio==4.0.0
awscli==1.29.49
botocore==1.31.49
certifi==2023.7.22
cftime==1.6.2
charset-normalizer==3.2.0
click==8.1.7
climate-assessment==0.1.3
colorama==0.4.4
contourpy==1.1.1
cycler==0.11.0
docutils==0.16
et-xmlfile==1.1.0
exceptiongroup==1.1.3
f90nml==1.4.3
fair==1.6.2
fonttools==4.42.1
globalwarmingpotentials==0.9.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
iam-units==2023.9.12
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.23.5
openpyxl==3.1.2
openscm-runner==0.9.1
openscm-units==0.5.0
packaging==23.1
pandas==1.5.3
pandas-datapackage-reader==0.18.0
Pillow==10.0.1
Pint==0.19.2
Pint-Pandas==0.4
platformdirs==3.10.0
ply==3.11
pooch==1.7.0
pyam-iamc==1.7.0
pyasn1==0.5.0
PyJWT==2.8.0
pymagicc==2.1.3
Pyomo==6.6.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.2
scipy==1.11.2
scmdata==0.15.0
seaborn==0.12.2
setuptools-scm==7.1.0
silicone==1.3.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.16
wquantiles==0.6
xarray==2023.8.0
xlrd==2.0.1
XlsxWriter==3.1.4
zipp==3.17.0
